requirements.txt
google-auth>=2.29.0
requests>=2.28.0
